# api/http/swagger.yaml
basePath: /api/v1
definitions:
  handlers.DeviceRequest:
    properties:
      password:
        description: |-
          Пароль пользователя
          required: true
          example: mySecret123
        type: string
      public_key:
        description: |-
          Публичный ключ устройства
          required: true
          example: MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAr...
        type: string
      username:
        description: |-
          Username пользователя
          required: true
          example: johndoe
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      device_uuid:
        description: |-
          UUID устройства
          required: true
          example: 4c2b87cd-8c44-4546-89df-7a751cbac96e
        type: string
      password:
        description: |-
          Пароль пользователя
          required: true
          example: mySecret123
        type: string
      username:
        description: |-
          Username пользователя
          required: true
          example: johndoe
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      password:
        description: |-
          Пароль пользователя
          required: true
          example: mySecret123
        type: string
      username:
        description: |-
          Username пользователя
          required: true
          example: johndoe
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для защищённого обмена сообщениями между пользователями
  title: bil-message API
  version: "1.0"
paths:
  /auth/device:
    post:
      consumes:
      - application/json
      description: Привязывает новое устройство к пользователю и возвращает UUID устройства
      parameters:
      - description: Данные устройства
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DeviceRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: UUID устройства
          schema:
            type: string
        "400":
          description: Неверные учетные данные или некорректные данные запроса
        "500":
          description: Внутренняя ошибка сервера
      summary: Добавление нового устройства
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет username, password и deviceUUID, возвращает JWT в заголовке
        Authorization
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: JWT токен успешно сгенерирован и возвращен в заголовке Authorization
        "400":
          description: Неверные учетные данные или некорректные данные запроса
        "500":
          description: Внутренняя ошибка сервера
      summary: Вход пользователя
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя с заданными username и password
      parameters:
      - description: Данные пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Пользователь успешно зарегистрирован
        "400":
          description: Некорректные данные запроса
        "409":
          description: Пользователь с таким именем уже существует
        "500":
          description: Внутренняя ошибка сервера
      summary: Регистрация нового пользователя
      tags:
      - Auth
  /chat:
    post:
      consumes:
      - text/plain
      description: Создаёт новую комнату для текущего пользователя
      produces:
      - text/plain
      responses:
        "200":
          description: UUID комнаты
          schema:
            type: string
        "400":
          description: Некорректные данные запроса
        "401":
          description: Неавторизован
        "500":
          description: Внутренняя ошибка сервера
      summary: Создание новой комнаты
      tags:
      - Chat
  /chat/{chat-uuid}:
    delete:
      consumes:
      - text/plain
      description: Удаляет комнату по UUID
      parameters:
      - description: UUID комнаты
        in: path
        name: chat-uuid
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Комната успешно удалена
        "400":
          description: Некорректные данные запроса
        "401":
          description: Неавторизован
        "404":
          description: Комната не найдена
        "500":
          description: Внутренняя ошибка сервера
      summary: Удаление комнаты
      tags:
      - Chat
  /chat/{chat-uuid}/{member-uuid}:
    delete:
      consumes:
      - text/plain
      description: Удаляет указанного пользователя (member-uuid) из комнаты
      parameters:
      - description: UUID комнаты
        in: path
        name: chat-uuid
        required: true
        type: string
      - description: UUID пользователя
        in: path
        name: member-uuid
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Пользователь успешно удалён
        "400":
          description: Некорректные данные запроса
        "401":
          description: Неавторизован
        "404":
          description: Комната или пользователь не найдены
        "500":
          description: Внутренняя ошибка сервера
      summary: Удаление пользователя из комнаты
      tags:
      - Chat
    post:
      consumes:
      - text/plain
      description: Добавляет указанного пользователя (member-uuid) в комнату
      parameters:
      - description: UUID комнаты
        in: path
        name: chat-uuid
        required: true
        type: string
      - description: UUID пользователя
        in: path
        name: member-uuid
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Пользователь успешно добавлен
        "400":
          description: Некорректные данные запроса
        "401":
          description: Неавторизован
        "404":
          description: Комната не найдена
        "500":
          description: Внутренняя ошибка сервера
      summary: Добавление пользователя в комнату
      tags:
      - Chat
swagger: "2.0"